pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "trader"
version = "0.1.0"
description = ""
authors = [
    "David Minarsch <david.minarsch@googlemail.com>",
    "Jose Moreira Sanchez <jose.moreira.sanchez@valory.xyz>",
    "Adamantios Zaras <adamantios.zaras@valory.xyz>",
]
readme = "README.md"
repository = "https://github.com/valory-xyz/trader"
keywords = [
    "trader",
    "autonomy",
    "autonomous",
    "autonolas",
    "service",
    "multi",
    "agent",
]
classifiers = [
    "Environment :: Console",
    "Environment :: Web Environment",
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: Microsoft",
    "Operating System :: Unix",
    "Programming Language :: Python :: 3.10",
    "Topic :: Communications",
    "Topic :: Internet",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development",
    "Topic :: System",
]
[[tool.poetry.packages]]
include = "packages"

[tool.poetry.dependencies]
python = "<3.12,>=3.9"
open-autonomy = "==0.19.9"
requests = "<2.31.2,>=2.28.1"
py-multibase = "==1.0.3"
py-multicodec = "==0.2.1"
py-eth-sig-utils = "*"
grpcio = "==1.53.0"
asn1crypto = "<1.5.0,>=1.4.0"
open-aea-ledger-ethereum = "==1.65.0"
open-aea-ledger-cosmos = "==1.65.0"
protobuf = "<4.25.0,>=4.21.6"
hypothesis = "==6.21.6"
open-aea-test-autonomy = "==0.19.9"
web3 = "<7,>=6.0.0"
ipfshttpclient = "==0.8.0a2"
open-aea-cli-ipfs = "==1.65.0"
aiohttp = "<4.0.0,>=3.8.5"
certifi = "==2021.10.8"
multidict = "*"
ecdsa = ">=0.15"
eth_typing = "*"
hexbytes = "==0.3.1"
packaging = "*"
py-ecc = "==6.0.0"
pytz = "==2022.2.1"
typing_extensions = "==4.13.2"
websocket_client = "<1,>=0.32.0"
toml = "==0.10.2"
pytest-asyncio = "*"
eth-utils = "==2.2.0"
eth-abi = "==4.0.0"
pycryptodome = "==3.18.0"
pytest = "==7.2.1"
urllib3 = "==1.26.16"
jsonschema = "<4.4.0,>=4.3.0"
openapi-core = "==0.15.0"
openapi-spec-validator = "<0.5.0,>=0.4.0"
pyinstaller = "==6.8.0"
google-generativeai = "0.8.2"

[tool.poetry.dependencies.tomte]
version = "==0.2.17"
extras = ["cli", "tests"]
